refactored packet management into multiple files
